+++ lib/regexp/index.ts
@@ -1,0 +1,11 @@
+'use strict';
+
+import {isRegExp} from './is';
+import {sanitizeRegExp} from './sanitize';
+
+export {
+    isRegExp,
+    isRegExp as is,
+    sanitizeRegExp,
+    sanitizeRegExp as sanitize,
+};